EFCoreMovie3_RazorPages - Solution and EFCoreTeaching_RazorPages - Solution finished

+++ EFCoreMovie3_RazorPages - Solution/Pages/Studios/Delete.cshtml.cs
@@ -11,20 +11,27 @@ namespace EFCoreMovie3_RazorPages_Solution.Pages.Studios
         public DeleteModel(IStudioService service)
         {
             this.service = service;
+            Studio = new Studio();
         }
         
            
         public Studio Studio { get; set; }
         public void OnGet(int id)
         {
-
-            Studio = service.GetStudioById(id);
+            Studio ? toBeDeleted = service.GetStudioById(id);
+            if (toBeDeleted != null)
+            {
+                Studio = toBeDeleted;
+            }
         }
 
         public IActionResult OnPost(int id)
         {
-            Studio toBeDeleted = service.GetStudioById(id);
-            service.DeleteStudio(toBeDeleted);
+            Studio ? toBeDeleted = service.GetStudioById(id);
+            if (toBeDeleted != null)
+            {
+                service.DeleteStudio(toBeDeleted);
+            }
             return RedirectToPage("GetStudios");
         }
     }
